on price focus out not executed on edit mode false
this eliminates the possibility of the user being on
product price +ivu entry, changing product from the list and then
triggering price focus out after accessing the other product.

because edit mode is false if form widgets are disabled, then on
displaying product details it will be false, so on price focus out
will not be triggered after accessing the other product.