fixed to sell report bug
when no previous excel file found, had an attribute error
AttributeError: 'NoneType' object has no attribute 'strftime'
when trying to input in the new excel the date of the previous excel.

fixed by adding a condition to evaluate if there is a previous excel,
add date of previous excel, if there's not, add a fixed text without date.