Update: products to sell report
placed all reports inside a new folder.

New folder is created and renamed to the date of the last report added.

Added a hyphen "-" at the beginning of the folder name to keep it
at the beginning of all the folders.

tested with no folders, with folder already created
with no files, with previous reports, with all reports.

Both products_to_sell_report and get_previous_excel_report_data
seems to be working as intended.