update: added limit to backups
Inventory Management document now backs up in a folder inside
Excel Backups called Inventory Management Backups and its limited
to 100 backups.

Products to Sell Reports Backups is inside the Excel Backups too, and
it is also limited to 100 backups.

If there's 200 backups, it uses a while to eliminate the oldest backups
by getting the backup names, sorting them so oldest one is first, and
then get the value of the first one of the list, which is the oldest
and then deletes it. it keeps doing that while the backups is
equal to 100 or less, which means that in 100 backups, it deletes
the 100th so 99 remains, and then the code continues to create the
new backup.

tested it with a lower limit and it worked as intended.
increased the limit to 100 on both cases again.