update: Added Rack ID entry widget
it reads the rack id from the excel for the selected product, and
it places the value in the rack id entry widget.

the program allows its modification and to save the contents of it
in case the user decides to change the rack id for the selected product.

everything seems to work as expected. no bugs found yet.