bugfix: folder rename
fixed a bug where folders used to be renamed with
double dash (-) rather than a single one after the product name.

also added a temporary fix to database append rather than
modify current data. db is not needed, it only
gets the folder names and paths on load and uses it
so I'm deleting the db on load so it erases the old
folder names and paths and updates it with the current one.

Can be done in json for efficiency and simplicity.